+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This version drops IRC Bot from Eventum Core, see #371
 - Fix corrupted note body on blocked multi-part emails (@balsdorf)
 - Drop IRC Bot from eventum core, available as eventum/irc-bot instead (@glensc, #371)
 - Fix parsing of link references in markdown (@glensc, #367)
+- Add Factory support for Extension to construct it's own classes (@glensc, #375)
 
 [3.5.0]: https://github.com/eventum/eventum/compare/v3.4.2...master